# [sw, otbn] Speed up ECDSA P-256 test