@vogelpi 2022-08-22
Thanks for your reviews. I've just discovered that a second, redundant zeroing of DMEM was done as part of the `otbn_load_app()` used in several tests. The otbn_smoketest also explicitly zeroed DMEM prior to performing the barrett384 multiplication. All this isn't needed and I've now removed that, too.

With that, the ECDSA P-256 test is even faster:
```console
I00001 otbn_ecdsa_op_irq_test.c:246] Initialization took 139585 cycles or 1395 us @ 100 MHz.
I00002 otbn_ecdsa_op_irq_test.c:406] Signing
I00003 otbn_ecdsa_op_irq_test.c:246] Sign took 9256 cycles or 92 us @ 100 MHz.
I00004 otbn_ecdsa_op_irq_test.c:412] Wiping OTBN DMEM and reloading app
I00005 otbn_ecdsa_op_irq_test.c:246] Wipe and Reload took 140140 cycles or 1401 us @ 100 MHz.
I00006 otbn_ecdsa_op_irq_test.c:421] Verifying
I00007 otbn_ecdsa_op_irq_test.c:246] Verify took 12912 cycles or 129 us @ 100 MHz.
I00008 otbn_ecdsa_op_irq_test.c:433] Wiping OTBN DMEM
I00009 otbn_ecdsa_op_irq_test.c:246] Wipe took 956 cycles or 9 us @ 100 MHz.
I00010 ottf_main.c:56] Finished sw/device/tests/otbn_ecdsa_op_irq_test.c
I00011 status.c:28] PASS!
```