@sriyerg 2022-08-22
Thank you very much @vogelpi for this work. Would you please also update the `run_timeout_mins` setting in `chip_sim_cfg.hjson` accordingly?