@vogelpi 2022-08-23
> Thank you very much @vogelpi for this work. Would you please also update the `run_timeout_mins` setting in `chip_sim_cfg.hjson` accordingly?

Hi @sriyerg ,
yes I was planning to do that. But I am not sure what timeout to choose. On our workstation, the latest version of the test still takes 3.5 hours:
```
|  Milestone  |      Name       | Tests                     |  Max Job Runtime  |  Simulated Time  |  Passing  |  Total  |  Pass Rate  |
|:-----------:|:---------------:|:--------------------------|:-----------------:|:----------------:|:---------:|:-------:|:-----------:|
|     V2      | chip_sw_otbn_op | chip_sw_otbn_ecdsa_op_irq |      3.510h       |     17.909ms     |     1     |    1    |  100.00 %   |
```
However, the latest nightly result on the dashboard is based on the previous version of the test, but it was still a lot faster:
```
|  Milestone  |      Name       | Tests                     |  Max Job Runtime  |  Simulated Time  |  Passing  |  Total  |  Pass Rate  |
|:-----------:|:---------------:|:--------------------------|:-----------------:|:----------------:|:---------:|:-------:|:-----------:|
|     V2      | chip_sw_otbn_op | chip_sw_otbn_ecdsa_op_irq |      1.141h       |     24.317ms     |     1     |    1    |  100.00 %   |
```
The previous timeout was set to 400min (= 6.6 hours). IIRC you reported a run time of roughly 5 hours last time. So, I would suggest to set something like 300min (= 5 hours) now. Does that make sense?