@vogelpi 2022-08-23
Okay, I'll prepare a PR for this. Thanks for your feedback @sriyerg .